chore(lint): enable max-classes-per-file
Split result variants into separate modules while preserving the existing result facade and type contract.

Closes #70
