+++ packages/cli/src/utils/result.ts
@@ -1,28 +1,2 @@
-interface Result<T, E> {
-  isOk: () => this is Ok<T, E>;
-  isErr: () => this is Err<T, E>;
-}
-
-export class Ok<T, E> implements Result<T, E> {
-  constructor(readonly value: T) {}
-
-  isOk(): this is Ok<T, E> {
-    return true;
-  }
-
-  isErr(): this is Err<T, E> {
-    return false;
-  }
-}
-
-export class Err<T, E> implements Result<T, E> {
-  constructor(readonly error: E) {}
-
-  isOk(): this is Ok<T, E> {
-    return false;
-  }
-
-  isErr(): this is Err<T, E> {
-    return true;
-  }
-}
+export { Err } from "./err";
+export { Ok } from "./ok";
